- Refactored based off prev commit messages
