+++ app/src/main/java/app/revanced/manager/ui/component/bundle/BundleTextContent.kt
@@ -2,10 +2,6 @@ package app.revanced.manager.ui.component.bundle
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Topic
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -17,15 +13,10 @@ import app.revanced.manager.R
 @Composable
 fun BundleTextContent(
     name: String,
-    isImportPage: Boolean = true,
     onNameChange: (String) -> Unit = {},
     isLocal: Boolean,
     remoteUrl: String,
     onRemoteUrlChange: (String) -> Unit = {},
-    patchBundleText: String,
-    onPatchLauncherClick: () -> Unit = {},
-    integrationText: String = "",
-    onIntegrationLauncherClick: () -> Unit = {},
 ) {
     OutlinedTextField(
         modifier = Modifier
@@ -48,47 +39,5 @@ fun BundleTextContent(
                 Text(stringResource(R.string.bundle_input_source_url))
             }
         )
-    } else {
-        if(isImportPage) {
-            OutlinedTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 16.dp),
-                value = patchBundleText,
-                onValueChange = {},
-                label = {
-                    Text("Patches Source File")
-                },
-                trailingIcon = {
-                    IconButton(
-                        onClick = onPatchLauncherClick
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Topic,
-                            contentDescription = null
-                        )
-                    }
-                }
-            )
-
-            OutlinedTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 16.dp),
-                value = integrationText,
-                onValueChange = {},
-                label = {
-                    Text("Integrations Source File")
-                },
-                trailingIcon = {
-                    IconButton(onClick = onIntegrationLauncherClick) {
-                        Icon(
-                            imageVector = Icons.Default.Topic,
-                            contentDescription = null
-                        )
-                    }
-                }
-            )
-        }
     }
 }
